# openapi.yaml
openapi: 3.0.3
info:
  title: Krawlet API
  version: 1.22.1
  description: |
    REST API for the Krawlet Minecraft economy tracking system.

    ## Authentication
    Most endpoints support optional API key authentication via Bearer token. 
    Include your API key in the `Authorization` header:
    ```
    Authorization: Bearer kraw_your_api_key_here
    ```

    ## Rate Limiting
    - Anonymous requests: 100 requests/hour
    - Free tier: 1,000 requests/hour
    - Premium tier: 5,000 requests/hour (Contact me with a valid reason to request access)
    - Service tiers (shopsync, enderstorage, worker, internal): Configurable per key

    Rate limit information is included in response headers:
    - `X-RateLimit-Limit`: Maximum requests per hour
    - `X-RateLimit-Remaining`: Requests remaining in current window
    - `X-RateLimit-Reset`: Unix timestamp when the limit resets

  contact:
    name: Krawlet API Support
    url: https://github.com/Twijn/krawlet-api
servers:
  - url: https://api.krawlet.cc
    description: Production API
  - url: http://localhost:3000/api
    description: Local development
security:
  - ApiKeyAuth: []
  - {}
...
paths:
...
  /v1/transfers/contents:
    get:
      tags: [Transfers]
      summary: Get live ender storage contents
      description: |
        Query a worker to list the current items in the authenticated player's ender storage.
        The worker sets the frequency, calls `list()`, and returns the result synchronously.
        Typically completes in under 30ms when a worker is available.

        **Concurrent worker cap:**
        - A Minecraft user can have at most 3 concurrent worker jobs total across transfer jobs and storage-content queries.
        - Requests above this limit return `429 TOO_MANY_WORKERS`.
        - This limit is globally enforced per Minecraft UUID and may vary by API key tier in the future.

        **Requires authentication:** This endpoint requires a valid API key with an associated
        Minecraft account that has ender storage color data configured.

        **Error conditions:**
        - `429` — User has reached the maximum concurrent worker jobs
        - `503` — No free worker is available; retry after a moment
        - `503` — Worker timed out (5s) or disconnected mid-query
        - `400` — API key has no associated Minecraft player or the player has no color data
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Current contents of the player's ender storage
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: '#/components/schemas/StorageSlotItem'
              example:
                success: true
                data:
                  items:
                    - name: 'minecraft:diamond'
                      count: 32
                    - name: 'minecraft:enchanted_book'
                      count: 1
                      nbt: '312db0d27bebe67a3084cbcff45c56cd'
                meta:
                  timestamp: '2026-04-12T10:30:00.000Z'
                  elapsed: 14
                  version: '1.0.0'
                  requestId: 'abc-123'
        '400':
          description: API key missing Minecraft player data or player missing color data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Player not found for this API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many concurrent worker jobs for this Minecraft user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: No worker available or worker timed out / disconnected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/transfers:
    get:
      tags: [Transfers]
      summary: List transfers
      description: |
        Retrieve all transfers where the authenticated player is the sender or recipient.

        **Requires authentication:** This endpoint requires a valid API key with an associated Minecraft account.
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: List of transfers
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Transfer'
              example:
                success: true
                data:
                  - id: 'a1b2c3d4-e5f6-7890-abcd-ef1234567890'
                    status: completed
                    error: null
                    fromUUID: 'd98440d6-5117-4ac8-bd50-70b086101e3e'
                    fromUsername: Twijn
                    toUUID: 'a1234567-89ab-cdef-0123-456789abcdef'
                    toUsername: Player2
                    itemName: 'minecraft:diamond'
                    itemNbt: null
                    quantity: 64
                    quantityTransferred: 64
                    timestamp: '2026-04-12T10:30:00.000Z'
                meta:
                  timestamp: '2026-04-12T10:30:00.000Z'
                  elapsed: 12
                  version: '1.0.0'
                  requestId: 'abc-123'
        '400':
          description: API key is missing Minecraft player data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [Transfers]
      summary: Queue a transfer
      x-codeSamples:
        - lang: lua
          label: Lua (CC:Tweaked) - Klog
          source: |
            local createKlog = require('klog') -- https://krawlet.cc/klog.lua

            local klog = createKlog('right', {
              apiKey = settings.get('klog.apiKey'),
              apiUrl = 'https://api.krawlet.cc/v1/'
            })

            local transfer, err = klog.transfer({
              to = 'Player2',
              itemName = 'minecraft:diamond',
              quantity = 64,
              pollInterval = 0.5
            })

            if transfer then
              print('Transfer completed! ID:', transfer.id)
            else
              printError('Transfer failed: ' .. err)
            end
      description: |
        Queue a new ender storage item transfer from the authenticated player to another player.
        The transfer will be executed by the next available worker.

        **Concurrent worker cap:**
        - A Minecraft user can have at most 3 concurrent worker jobs total across transfer jobs and storage-content queries.
        - Requests above this limit return `429 TOO_MANY_WORKERS`.
        - This limit is globally enforced per Minecraft UUID and may vary by API key tier in the future.

        **Requires authentication:** This endpoint requires a valid API key with an associated Minecraft account.

        **Filters:**
        - `itemName` — only transfer items matching this Minecraft item ID
        - `itemNbt` — only transfer items matching this NBT hash string (can be combined with `itemName`)
        - `quantity` — stop after moving this many items total
        - `timeout` — seconds to wait for items/space before failing (0.5–30, default 5)
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferCreateRequest'
            examples:
              allItems:
                summary: Transfer all items
                value:
                  to: Player2
              byName:
                summary: Transfer specific item by name
                value:
                  to: Player2
                  itemName: 'minecraft:diamond'
                  quantity: 64
              byNbt:
                summary: Transfer items matching name + NBT hash
                value:
                  to: Player2
                  itemName: 'minecraft:enchanted_book'
                  itemNbt: '312db0d27bebe67a3084cbcff45c56cd'
                  quantity: 1
      responses:
        '200':
          description: Transfer queued successfully
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Transfer'
              example:
                success: true
                data:
                  id: 'a1b2c3d4-e5f6-7890-abcd-ef1234567890'
                  status: pending
                  error: null
                  fromUUID: 'd98440d6-5117-4ac8-bd50-70b086101e3e'
                  fromUsername: Twijn
                  toUUID: 'a1234567-89ab-cdef-0123-456789abcdef'
                  toUsername: Player2
                  itemName: 'minecraft:diamond'
                  itemNbt: null
                  quantity: 64
                  quantityTransferred: 0
                  timestamp: '2026-04-12T10:30:00.000Z'
                meta:
                  timestamp: '2026-04-12T10:30:00.000Z'
                  elapsed: 18
                  version: '1.0.0'
                  requestId: 'def-456'
        '400':
          description: Invalid payload or player data missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many concurrent worker jobs for this Minecraft user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/transfers/{transferId}:
    get:
      tags: [Transfers]
      summary: Get transfer by ID
      description: |
        Retrieve details for a specific transfer. Only accessible by the sender or recipient.

        **Requires authentication:** This endpoint requires a valid API key with an associated Minecraft account.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: transferId
          in: path
          required: true
          description: Transfer UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transfer details
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Transfer'
        '403':
          description: You do not have permission to access this transfer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Transfer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/transfers/{transferId}/cancel:
    post:
      tags: [Transfers]
      summary: Cancel a transfer
      x-codeSamples:
        - lang: lua
          label: Lua (CC:Tweaked) - Klog
          source: |
            local createKlog = require('klog') -- https://krawlet.cc/klog.lua

            local klog = createKlog('right', {
              apiKey = settings.get('klog.apiKey'),
              apiUrl = 'https://api.krawlet.cc/v1/'
            })

            local ok, err = klog.cancelTransfer('a1b2c3d4-e5f6-7890-abcd-ef1234567890')
            if not ok then
              printError('Cancel failed: ' .. err)
            end
      description: |
        Request cancellation of a pending or in-progress transfer. Only the sender can cancel.
        If a worker is actively processing the transfer, a cancel signal is sent and the transfer
        is finalized as `cancelled` once acknowledged.

        **Requires authentication:** This endpoint requires a valid API key with an associated Minecraft account.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: transferId
          in: path
          required: true
          description: Transfer UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transfer cancelled successfully
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Transfer'
              example:
                success: true
                data:
                  id: 'a1b2c3d4-e5f6-7890-abcd-ef1234567890'
                  status: cancelled
                  error: null
                  fromUUID: 'd98440d6-5117-4ac8-bd50-70b086101e3e'
                  fromUsername: Twijn
                  toUUID: 'a1234567-89ab-cdef-0123-456789abcdef'
                  toUsername: Player2
                  itemName: 'minecraft:diamond'
                  itemNbt: null
                  quantity: 64
                  quantityTransferred: 32
                  timestamp: '2026-04-12T10:30:00.000Z'
                meta:
                  timestamp: '2026-04-12T10:30:00.000Z'
                  elapsed: 10
                  version: '1.0.0'
                  requestId: 'ghi-789'
        '400':
          description: Transfer cannot be cancelled (already completed/failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to cancel this transfer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Transfer not found or cannot be cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  headers:
    X-RateLimit-Limit:
      description: Maximum requests allowed in the current window.
      schema:
        type: integer
      example: 1000
    X-RateLimit-Remaining:
      description: Remaining requests in the current window.
      schema:
        type: integer
      example: 999
    X-RateLimit-Reset:
      description: Unix timestamp when the current rate-limit window resets.
      schema:
        type: integer
      example: 1736686200
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: 'API key authentication. Format: Bearer kraw_your_key_here'

    ShopSyncAuth:
      type: http
      scheme: bearer
      description: ShopSync authentication token

    EnderStorageAuth:
      type: http
      scheme: bearer
      description: Ender storage authentication token
  schemas:
    SuccessResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: Response data (varies by endpoint)
        meta:
          type: object
          required:
            - timestamp
            - elapsed
            - version
            - requestId
          properties:
            timestamp:
              type: string
              format: date-time
              description: Server timestamp
            elapsed:
              type: integer
              description: Request processing time in milliseconds
            version:
              type: string
              description: API version
              example: '1.0.0'
            requestId:
              type: string
              format: uuid
              description: Unique request identifier
            rateLimit:
              type: object
              description: Rate limit information (when authenticated)
              properties:
                limit:
                  type: integer
                  description: Maximum requests per hour
                remaining:
                  type: integer
                  description: Remaining requests in window
                reset:
                  type: integer
                  description: Unix timestamp when limit resets
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - meta
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - INVALID_API_KEY
                - RATE_LIMIT_EXCEEDED
                - TOO_MANY_WORKERS
                - NOT_FOUND
                - INTERNAL_ERROR
                - BAD_REQUEST
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error details
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            elapsed:
              type: integer
            version:
              type: string
            requestId:
              type: string
              format: uuid
...
    StorageSlotItem:
      type: object
      description: A single item stack in an ender storage inventory slot
      properties:
        name:
          type: string
          description: Minecraft item ID
          example: 'minecraft:diamond'
        count:
          type: integer
          description: Number of items in this stack
          example: 32
        nbt:
          type: string
          nullable: true
          description: Opaque hash string of the item's NBT data (present only for NBT-bearing items)
          example: '312db0d27bebe67a3084cbcff45c56cd'
    Transfer:
      type: object
      description: An ender storage item transfer between two players
      properties:
        id:
          type: string
          format: uuid
          description: Unique transfer ID
        status:
          type: string
          enum: [pending, in_progress, completed, failed, cancelled]
          description: Current status of the transfer
        error:
          type: string
          nullable: true
          description: Error message if the transfer failed
        workerId:
          type: integer
          nullable: true
          description: ID of the worker computer processing the transfer
        fromUUID:
          type: string
          format: uuid
          description: Sender's Minecraft UUID
        fromUsername:
          type: string
          description: Sender's Minecraft username
        toUUID:
          type: string
          format: uuid
          description: Recipient's Minecraft UUID
        toUsername:
          type: string
          description: Recipient's Minecraft username
        itemName:
          type: string
          nullable: true
          description: Minecraft item ID filter — only items with this name are transferred
          example: 'minecraft:diamond'
        itemNbt:
          type: string
          nullable: true
          description: NBT filter — only items matching this NBT hash string are transferred (can be combined with itemName)
          example: '312db0d27bebe67a3084cbcff45c56cd'
        quantity:
          type: integer
          nullable: true
          description: Maximum number of items to transfer (null means transfer all matching items)
          minimum: 1
        quantityTransferred:
          type: integer
          description: Number of items transferred so far
        timestamp:
          type: string
          format: date-time
          description: When the transfer was created
    TransferCreateRequest:
      type: object
      required:
        - to
      properties:
        to:
          type: string
          description: Recipient Minecraft username or UUID
          example: Player2
        itemName:
          type: string
          description: Only transfer items matching this Minecraft item ID
          example: 'minecraft:diamond'
        itemNbt:
          type: string
          description: Only transfer items matching this NBT hash string (can be combined with itemName)
          example: '312db0d27bebe67a3084cbcff45c56cd'
        quantity:
          type: integer
          description: Maximum number of items to transfer (omit to transfer all matching items)
          minimum: 1
          example: 64
        timeout:
          type: number
          description: Seconds to wait for items/space before failing (0.5–30, default 5)
          minimum: 0.5
          maximum: 30
          example: 5